Wrote methods and did some debugging
In short, the algorithm is working despite being held together by popsicle sticks and rubber bands.

At length and in no particular order:

I've changed the behavior of ParseCSV (when you give it a pre-generated list of attributes) to replace unidentified values when they occur. While it will create multiple fractional counts for multiple missing values, it will only fill in one missing value per fractional count, and will not fill the rest in. On the other hand, it shouldn't go deleting extra cases when the same ones are marked multiple times as missing values  indata when it tries to replace them.

Created function FindTestError() which takes in a dataset, a tree, and a set of attributes that will return the error of the tree's performance on the dataset as a double. Neat.

Fixed a bunch of bugs in the ID3 algorithm

Made a recursive function that returns a string that depicts the entire tree. Nicely formatted if I do say so myself.

Figured out how to do the relative file paths. Check the three test cases I've written up to see it.

I think that's it. I guess I also made a file to put all of the homework testing in with a Main function. So yeah.