Assignment Stuff and Bugfixing
Did the car tests and excountered half a million bugs. aaaahh. That's not good. Anyhow, I fixed most of them.

The main one I'm aware of that remains is that only Majority error seems to work for building the tree. So that's a thing.

Unfortunately, the assignment is now late. \_:)_/. I guess I'll just need to shoot for the 80%.