ID3 Algorithm, to my understanding, works
Fixed a problem that causes the proportion calculations to fail. As such, the algorithm actually works. However, all three entropy calculations appear to generate the same trees most of the time, or all of the time. Can't confirm that bit.

Finished question one in part two of the homework (code for that is "part1"  in main). Tests the cars dataset.