Attributes can now be numeric
Numerical attributes are binary attributes that are either above/equal to or below the median data point for that attribute in the training data.

On a side note, the testing document seems to have ceased function somewhere along the way.

Changed get label distribution to no longer accept an attribute number as  the data is included in the attribute that's passed in. Protects me from myself.

Put in the majority value filling for missing items, and split that and the fractional counts into their own methods that get called by ParseCSV.

Put in the code that Sets up the data for building the bank trees, and builds the trees. It's done! At long last. run.sh is also included in the decision tree folder.