Updated attribute and corresponding functions
Attributes now have a public readonly enum (replacing the numeric bool) that corresponds to whether it is binary numeric, numeric, or categorical

Updated Functions:

Constructor:
-Saw that one coming

GetVarID(String Variant);
-Uses a switch.
-Now returns a double (helpful for pure numberic attributes)
-Being able to return anything would actually be useful. Too bad this isnt't Python.

Updated Cases:

AttributeVals is now an array of doubles

Updated ID3:

Now throws exceptions when you try to use pure numeric attributes on it. Also casts attributevals to be integers when using it and assumes that the attributes are not purely numerical.