Implemented the Perceptron algorithm
... but I'm not sure if it's functional. The voted perceptron works and that gives me confidence.

Anyhow, the perceptron algorithm is implemented like the linear regression learner where you put all the stuff into an object and call a method on it to run the algorithm.

There are four variants of the learner which all use the same methods, but will end with different results. This is based on the PType enum that gets passed.

There is the normal perceptron, voted perceptron,  averaged perceptron, and margin perceptron.

Note: Margin perceptron takes an extra argument to specify the margin to check for on correct predictions. Defaults to 1 if there is no extra input.

Cheers.