Made an attempt at SVM. Incomplete.
I've implemented what I thought to be SVM SubGradient decent, but it would seem that I do not understand it at all.

The program does nicely print out its results to the console when run.

Change to class "Case" in General_Tools
-Implemented ColXtoY()
Method takes in a List of cases and swaps all instances of double X with Y in the specified Collumn.